+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Changed
+- `TypeSafeKey.Simple` now has a default `#apply` implementation
 
 ## [3.0.1] - 2018-05-13
 ### Changed