+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+No unreleased changes
+
+## [3.0.2] - 2018-05-21
 ### Added
 - `IterableLens#mapping`, an `Iso` that maps values
 
@@ -323,7 +326,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `Monadic/Dyadic/TriadicFunction`, `Predicate`, `Tuple2`, `Tuple3`
 - `Functor`, `BiFunctor`, `ProFunctor` 
 
-[Unreleased]: https://github.com/palatable/lambda/compare/lambda-3.0.1...HEAD
+[Unreleased]: https://github.com/palatable/lambda/compare/lambda-3.0.2...HEAD
+[3.0.2]: https://github.com/palatable/lambda/compare/lambda-3.0.1...lambda-3.0.2
 [3.0.1]: https://github.com/palatable/lambda/compare/lambda-3.0.0...lambda-3.0.1
 [3.0.0]: https://github.com/palatable/lambda/compare/lambda-2.1.1...lambda-3.0.0
 [2.1.1]: https://github.com/palatable/lambda/compare/lambda-2.1.0...lambda-2.1.1